refactor(spec): lift verify/forward_* to BaseSpecWorker (fold #1080 review (1))
verify(), forward_target_extend(), forward_batch_speculative_generation()
only touch BaseSpecWorker state (target_worker, draft_worker, mesh,
speculative_num_*); moving them up makes EAGLEWorker and
MultiLayerEAGLEWorker thin draft_worker-injection wrappers.

BaseDraftWorker gains explicit abstract draft_extend_for_{prefill,decode}
+ draft_model_runner so the contract is visible.

(3) precompile multi-layer warm-up left as TODO (currently
--disable-precompile in E2E; only layer 0 would be warmed via
_worker delegation).
