+++ python/sgl_jax/srt/speculative/eagle_worker.py
@@ -4,23 +4,16 @@ import time
 
 import jax
 import jax.numpy as jnp
-import numpy as np
 from tqdm import tqdm
 
 from sgl_jax.srt.layers.logits_processor import LogitsProcessorOutput
 from sgl_jax.srt.layers.sampler import get_token_ids_logprobs, get_top_logprobs
-from sgl_jax.srt.managers.schedule_batch import ModelWorkerBatch, ScheduleBatch
-from sgl_jax.srt.managers.scheduler import GenerationBatchResult
+from sgl_jax.srt.managers.schedule_batch import ScheduleBatch
 from sgl_jax.srt.managers.tp_worker import ModelWorker
 from sgl_jax.srt.model_executor.forward_batch_info import CaptureHiddenMode, ForwardMode
-from sgl_jax.srt.sampling.sampling_batch_info import SamplingMetadata
-from sgl_jax.srt.speculative.base_worker import BaseSpecWorker, replicate_to_mesh
+from sgl_jax.srt.speculative.base_worker import BaseSpecWorker
 from sgl_jax.srt.speculative.eagle_draft_worker import EagleDraftWorker
-from sgl_jax.srt.speculative.eagle_util import (
-    EagleDraftInput,
-    EagleVerifyInput,
-    EagleVerifyOutput,
-)
+from sgl_jax.srt.speculative.eagle_util import EagleDraftInput, EagleVerifyOutput
 from sgl_jax.srt.utils.common_utils import get_bool_env_var
 
 logger = logging.getLogger(__name__)
@@ -36,172 +29,14 @@ class EAGLEWorker(BaseSpecWorker):
     """
 
     def __init__(self, server_args, target_worker: ModelWorker, draft_worker=None):
-        self.server_args = server_args
-        self._target_worker = target_worker
-        self._draft_worker = draft_worker or EagleDraftWorker(server_args, target_worker)
-
-        self.topk = server_args.speculative_eagle_topk
-        self.speculative_num_steps = server_args.speculative_num_steps
-        self.speculative_num_draft_tokens = server_args.speculative_num_draft_tokens
-        self.page_size = server_args.page_size
-        self.mesh = target_worker.mesh
-
-        from sgl_jax.srt.speculative.spec_info import SpeculativeAlgorithm
-
-        self.speculative_algorithm = SpeculativeAlgorithm.from_string(
-            server_args.speculative_algorithm
+        super().__init__(
+            server_args,
+            target_worker,
+            draft_worker or EagleDraftWorker(server_args, target_worker),
         )
 
-        self.req_to_token_pool, self.token_to_kv_pool_allocator = target_worker.get_memory_pool()
-
-        (
-            precompile_token_paddings,
-            precompile_bs_paddings,
-            precompile_cache_loc_paddings,
-        ) = target_worker.get_precompile_paddings()
-        self.precompile_bs_paddings = precompile_bs_paddings
-        self.precompile_cache_loc_paddings = precompile_cache_loc_paddings
-        self.precompile_token_paddings = precompile_token_paddings
-
-    # -- BaseSpecWorker interface --
-
-    @property
-    def target_worker(self) -> ModelWorker:
-        return self._target_worker
-
-    @property
-    def draft_worker(self) -> EagleDraftWorker:
-        return self._draft_worker
-
-    def forward_batch_speculative_generation(
-        self,
-        model_worker_batch: ModelWorkerBatch,
-    ):
-        if model_worker_batch.forward_mode.is_extend():
-            # FIXME(pc) add padding logic here
-            if model_worker_batch.sampling_info.temperatures.ndim == 1:
-                model_worker_batch.sampling_info.temperatures = (
-                    model_worker_batch.sampling_info.temperatures[:, None]
-                )
-            sampling_metadata = SamplingMetadata.from_model_worker_batch(
-                model_worker_batch,
-                len(model_worker_batch.seq_lens) - model_worker_batch.real_bs,
-                self.mesh,
-                vocab_size=self.target_worker.model_config.vocab_size,
-            )
-            # target extend
-            logits_output, next_token_ids, cache_miss_count, bid, seq_lens = (
-                self.forward_target_extend(model_worker_batch, sampling_metadata)
-            )
-            # draft extend for Update Draft State
-            self.draft_worker.draft_extend_for_prefill(
-                model_worker_batch, logits_output.hidden_states, next_token_ids
-            )
-            # FIXME(pc) refactor this to batch output
-            batch_output = GenerationBatchResult(
-                logits_output=logits_output,
-                next_token_ids=next_token_ids,
-                next_draft_input=model_worker_batch.spec_info,
-                allocate_lens=model_worker_batch.seq_lens[: model_worker_batch.real_bs],
-                bid=bid,
-                cache_miss_count=cache_miss_count,
-                extend_input_len_per_req=None,
-                extend_logprob_start_len_per_req=None,
-            )
-            return batch_output
-
-        else:
-            cur_allocate_lens = model_worker_batch.spec_info.allocate_lens
-            self.draft_worker.draft(model_worker_batch)
-
-            batch_output = self.verify(model_worker_batch, cur_allocate_lens)
-
-            self.draft_worker.draft_extend_for_decode(model_worker_batch, batch_output)
-
-            return batch_output
-
-    # -- Target model methods --
-
-    def forward_target_extend(
-        self, model_worker_batch: ModelWorkerBatch, sample_meta_data: SamplingMetadata
-    ) -> tuple[LogitsProcessorOutput, jax.Array, int, int, np.ndarray]:
-        model_worker_batch.capture_hidden_mode = CaptureHiddenMode.FULL
-        logits_output, next_token_ids, cache_miss_count = (
-            self.target_worker.forward_batch_generation(
-                model_worker_batch, sampling_metadata=sample_meta_data
-            )
-        )
-        return (
-            logits_output,
-            next_token_ids,
-            cache_miss_count,
-            model_worker_batch.bid,
-            model_worker_batch.seq_lens,
-        )
-
-    # -- Verify --
-
-    def verify(self, model_worker_batch: ModelWorkerBatch, cur_allocate_lens: jax.Array):
-        spec_info: EagleVerifyInput = model_worker_batch.spec_info
-        spec_info.allocate_lens = cur_allocate_lens
-        spec_info.prepare_for_verify(model_worker_batch, self.page_size, self.target_worker)
-        forward_metadata = self.target_worker.model_runner.attn_backend.get_eagle_forward_metadata(
-            model_worker_batch
-        )
-
-        logits_output, _, cache_miss_count = self.target_worker.forward_batch_generation(
-            model_worker_batch, skip_sample=True, forward_metadata=forward_metadata
-        )
-        logits_output.next_token_logits, logits_output.hidden_states = replicate_to_mesh(
-            self.mesh, logits_output.next_token_logits, logits_output.hidden_states
-        )
-        spec_info.hidden_states = logits_output.hidden_states
-        (
-            predict,
-            verified_id,
-            accept_length,
-            accept_index,
-        ) = spec_info.sample(
-            model_worker_batch,
-            logits_output,
-            self.draft_worker.draft_model_runner.rngs,
-            self.mesh,
-        )
-        # accept_index uses -1 for rejected slots; gathering with -1 picks the
-        # global last element, so dext later writes rejected tokens' draft-KV at
-        # a foreign position inside each req's page (corrupts prefix KV for all
-        # but the last req at bs>1). Redirect -1 to each req's own last slot.
-        # accept_index has length bs*(spec_steps+1); the gathered tensors have
-        # length bs*draft_token_num — equal at topk=1, distinct at topk>1.
-        draft_n = self.speculative_num_draft_tokens
-        accept_width = self.speculative_num_steps + 1
-        req_ids = np.arange(len(accept_index)) // accept_width
-        per_req_last = req_ids * draft_n + draft_n - 1
-        safe_index = np.where(accept_index >= 0, accept_index, per_req_last)
-        logits_output.next_token_logits = logits_output.next_token_logits[safe_index, :]
-        logits_output.hidden_states = logits_output.hidden_states[safe_index, :]
-        model_worker_batch.positions = model_worker_batch.positions[safe_index]
-        new_seq_lens = model_worker_batch.seq_lens + accept_length
-        next_draft_input = EagleDraftInput(
-            verified_id=verified_id,
-            new_seq_lens=new_seq_lens,
-            allocate_lens=cur_allocate_lens,
-            hidden_states=logits_output.hidden_states,
-        )
-
-        model_worker_batch.spec_info = next_draft_input
-        return GenerationBatchResult(
-            logits_output=logits_output,
-            next_token_ids=predict,
-            next_draft_input=next_draft_input,
-            accept_lens=accept_length,
-            # FIXME(pc) this field is for overlap
-            allocate_lens=cur_allocate_lens,
-            bid=model_worker_batch.bid,
-            cache_miss_count=cache_miss_count,
-            extend_input_len_per_req=None,
-            extend_logprob_start_len_per_req=None,
-        )
+    # -- BaseSpecWorker provides target_worker/draft_worker/verify/
+    #    forward_target_extend/forward_batch_speculative_generation --
 
     # -- Logprob post-processing --
